Delay sending mode change to wear os
Noticed that wear os won't always vibrate on receiving mode change
and that started happening when delay was added to the receiving
service before vibrating/showing toast because mode would change
frequently, when wifi dropped for example.
Thinking it might be that background service exits because it's not
aware of the coroutine being run.
Moving that delay to be at the source instead now